user logout and restricted pages

+++ base/views.py
@@ -1,13 +1,29 @@
 from django.shortcuts import render, redirect
-from .models import Room
+from base.models import *
+from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.db.models import Q
 from base.forms import *
 # Create your views here.
 
 
 def home(request):
-    rooms = Room.objects.all()
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) |
+                                Q(name__icontains=q) |
+                                Q(description__icontains=q)
+                                )
+
+    topics = Topic.objects.all()
+
+    room_count = rooms.count()
     context = {
-        'rooms': rooms
+        'rooms': rooms,
+        'topics': topics,
+        'room_count': room_count,
     }
     return render(request, 'base/home.html', context)
 
@@ -15,12 +31,13 @@ def home(request):
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {
-        'room': room
+        'room': room,
     }
 
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -35,9 +52,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('your are not allowed to update this room')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -51,10 +72,47 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('your are not allowed to update this room')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
 
     return render(request, 'base/delete.html', {'obj': room})
+
+
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid credentials')
+
+    context = {
+
+    }
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
